+++ Hero.h
