+++ main.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include "Game.h"
 #include "Monster.h"
-#include "Player.h"
+#include "Hero.h"
 #include <fstream>
 
 using namespace std;
@@ -17,15 +17,15 @@ int isNumeric(const string & chars) {
 
 int main (int argc, char *argv[]) {
 	
-    Player ent1(150, 10, "Maple", 15.2f);
-    Player ent2(45, 30, "Sally", 10.5f);
+    Hero ent1(150, 10, "Maple", 15.2f);
+    Hero ent2(45, 30, "Sally", 10.5f);
 	
     if(argc != 3){
         std::cout << "Wrong initial data given, using default values." << std::endl;
     }else{
         try{
-            ent1 = Player::parseUnit(argv[1]);
-            ent2 = Player::parseUnit(argv[2]);
+            ent1 = Hero::parseUnit(argv[1]);
+            ent2 = Hero::parseUnit(argv[2]);
         }catch(std::exception& e){
             std::cout << "An error occured while the program was running: " << e.what() << std::endl;
             return 1;
